requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-js-asset==2.0.0
django-mptt==0.13.4
Pillow==9.1.1
sqlparse==0.4.2
